Move enum duplicates removing to separate file
Functionality to remove duplicates from list of enums was moved to
separate file to allow both gds_dictgen.py and codegen.py to use it
without duplicating code. Also a few typos were spotted and fixed.
